Create optim.composite.ttest.r
This program calculates the relative power over varying available sample sizes under the condition that a finite number of individuals and a finite number of assays are available. In this case, composite sampling can aggregate some number of samples to allow enrollment of additional subjects. In almost all scenarios, optimal conditions can be determined with mild arithmetic, but identifying sufficient conditions may be necessary to account for study designs.